[Android] fix oom while uploading big files
For files larger than `Int.MAX_VALUE` the size wouldn't be returned
correctly. We have changed IPC to use `Long` for numbers instead. There
is a danger of returning a value larger than JS
`Number.MAX_SAFE_INTEGER` but it should cover most of our cases.

Kotlin's `readBytes()` takes `InputStream#available()` into account
which is not implemented for `BoundedInputStream` in a way that makes it
usable for us. After certain point it would return the rest of the file
instead of the limited size which led to `readBytes()` trying to
allocate a large slab of memory at once, which would fail with OOM.

We reimplemented `readBytes()` without taking `available()` into account
and we do not use `available()` For determining if the stream is empty.

+++ app-android/calendar/src/main/java/de/tutao/calendar/MainActivity.kt
@@ -163,12 +163,12 @@ class MainActivity : FragmentActivity(), ActivityUtils {
 				NetworkUtils.defaultClient,
 				{ fileId, bytes ->
 					lifecycleScope.launch {
-						commonNativeFacade.downloadProgress(fileId, bytes)
+						commonNativeFacade.downloadProgress(fileId, bytes.toLong())
 					}
 				},
 				{ fileId, bytes ->
 					lifecycleScope.launch {
-						commonNativeFacade.uploadProgress(fileId, bytes)
+						commonNativeFacade.uploadProgress(fileId, bytes.toLong())
 					}
 				}, "unused"
 			)
@@ -274,7 +274,7 @@ class MainActivity : FragmentActivity(), ActivityUtils {
 			)
 			val imeHeight = windowInsets.getInsets(ime()).bottom
 			lifecycleScope.launch {
-				mobileFacade.keyboardSizeChanged(imeHeight.toDp())
+				mobileFacade.keyboardSizeChanged(imeHeight.toDp().toLong())
 			}
 
 			// Convert raw pixels to density independent pixels
